requirements.txt
fastapi
uvicorn[standard]
pillow
python-dotenv
torch
numpy
sentence-transformers
scikit-learn